requirements.txt
# Core dependencies
requests>=2.31.0
fastapi>=0.104.1
uvicorn[standard]>=0.24.0
pydantic>=2.0.0
python-dotenv>=1.0.0
aiohttp>=3.8.0

# Strands Framework (REQUIRED for agents to work)
strands-agents[otel]>=0.1.0
strands-agents-tools>=0.1.0

# OpenTelemetry (for observability)
opentelemetry-api>=1.36.0
opentelemetry-sdk>=1.36.0
opentelemetry-exporter-otlp-proto-http>=1.36.0
opentelemetry-instrumentation-requests>=0.44b0
opentelemetry-instrumentation-urllib3>=0.44b0
opentelemetry-instrumentation-fastapi>=0.44b0

# Langfuse (for observability backend) - CRITICAL for tracing!
langfuse>=2.0.0

# Development (optional)
ruff>=0.1.6
black>=23.0.0
mypy>=1.0.0
